Additional modifications for x64
 * Use ptrdiff_t for the return type of onig_match and onig_search.
   'long' is not a 64-bit type in VC++.
 * Add casts to suppress unnecessary warnings.
 * Use Oniguruma-style cast: (type )val

+++ enc/utf32_le.c
@@ -152,7 +152,7 @@ utf32le_left_adjust_char_head(const UChar* start, const UChar* s)
 
   if (s <= start) return (UChar* )s;
 
-  rem = (s - start) % 4;
+  rem = (int )((s - start) % 4);
   return (UChar* )(s - rem);
 }
